Swap function variable from unsigned to signed
A function argument in InsertPrecipitatePhases could possibly take negative
values, but it was being assigned as an unsigned variable, resulting in
underflow.  Fixed by changing it back to a signed integer.

+++ Source/Plugins/SyntheticBuilding/SyntheticBuildingFilters/InsertPrecipitatePhases.cpp
@@ -1284,7 +1284,7 @@ void InsertPrecipitatePhases::move_precipitate(int32_t gnum, float xc, float yc,
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-void InsertPrecipitatePhases::update_exclusionZones(size_t gadd, int32_t gremove, Int32ArrayType::Pointer exclusionZonesPtr)
+void InsertPrecipitatePhases::update_exclusionZones(int32_t gadd, int32_t gremove, Int32ArrayType::Pointer exclusionZonesPtr)
 {
   size_t featureOwnersIdx = 0;
   int32_t* exclusionZones = exclusionZonesPtr->getPointer(0);
